bedtools: close coverage/map/closest/jaccard/groupby tails + 4 new subcommands
Drive the bedtools tail toward 100%, each validated byte-for-byte against the
real upstream bedtools binary (live-oracle tests, no goldens):

- bedcoverage: -abam/-ibam BAM-for-A wiring, -sorted (no-op), the verbatim
  mutually-exclusive-modes error, and float32 covered-fraction rounding.
- bedmap: BED/VCF/GFF/BAM auto-detect for -b, -c column-range + ops/cols
  mismatch errors verbatim, non-numeric warnings, -header/-split/-g/-prec,
  setprecision(10) numeric formatting (test-map.sh 57/57).
- bedclosest: faithful CloseSweep port — unsigned -d vs signed -D, per-record
  null shapes, opt-in distance column, -k/-io/-header, multi-db ordering
  (test-closest.sh 71/71 + kclosest 45/45).
- bedjaccard: -bed flag for BAM input (table emission parity).
- bedgroupby: setprecision(10) scalar-numeric formatting + --prec (t20 large
  ints -> scientific).
- NEW subcommands (previously missing): bedrandom (bedtools random, MT19937_64
  RNG byte-exact at -seed), bedshift (bedtools shift), bedunionbedg
  (bedtools unionbedg), bedoverlap (bedtools overlap) — full flag surface,
  live-upstream parity.

(bedmerge/bedintersect/bedgetfasta/bamtobed work from the same wave was
incomplete when agents hit the session limit; excluded here, to be redone
cleanly.)

+++ tools/bedcoverage/pkg/bedcoverage/bedcoverage.go
@@ -580,8 +580,15 @@ func joinTrailingComma(vs []int) string {
 
 // formatFraction prints the fraction column using 7 fixed decimals, matching
 // upstream `bedtools coverage` (e.g. "1.0000000", "0.7600000").
+//
+// Upstream computes the covered-fraction as a 32-bit float (the
+// numerator/denominator division happens in float arithmetic in
+// coverageFile.cpp / RecordOutputMgr) and prints it with 7 decimals, so the
+// last digit carries float32 rounding. For example 7/19 prints as
+// "0.3684210", not the float64-rounded "0.3684211". Narrow to float32 before
+// formatting to reproduce upstream byte-for-byte.
 func formatFraction(v float64) string {
-	return strconv.FormatFloat(v, 'f', 7, 64)
+	return strconv.FormatFloat(float64(float32(v)), 'f', 7, 64)
 }
 
 // formatFloatLoose prints a number with up to 7 significant digits, trimming
